Module refactoring
Move common random generation methods to one file

Make random methods exported, so they can be used in acceptance

Cover panic raised in `initClient` during acceptance tests
